refactored test helper classes into maven main artifact of grakn-test module
