+++ grakn-test/src/main/java/ai/grakn/test/GraphContext.java
@@ -27,10 +27,10 @@ import org.junit.runners.model.Statement;
 
 import java.util.function.Consumer;
 
-import static ai.grakn.graphs.TestGraph.loadFromFile;
 import static ai.grakn.test.GraknTestEnv.ensureCassandraRunning;
 import static ai.grakn.test.GraknTestEnv.randomKeyspace;
 import static ai.grakn.test.GraknTestEnv.usingTinker;
+import static ai.grakn.test.graphs.TestGraph.loadFromFile;
 
 /**
  *
