+++ grakn-test/src/test/java/ai/grakn/graphs/AdmissionsGraph.java
@@ -22,6 +22,7 @@ import ai.grakn.GraknGraph;
 import ai.grakn.concept.EntityType;
 import ai.grakn.concept.Instance;
 import ai.grakn.concept.ResourceType;
+import ai.grakn.test.graphs.TestGraph;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
