+++ grakn-test/src/test/java/ai/grakn/graphs/GeoGraph.java
@@ -26,6 +26,7 @@ import ai.grakn.concept.ResourceType;
 import ai.grakn.concept.RoleType;
 import ai.grakn.concept.RuleType;
 import ai.grakn.graql.Pattern;
+import ai.grakn.test.graphs.TestGraph;
 
 import java.util.function.Consumer;
