+++ grakn-test/src/test/java/ai/grakn/graphs/MatrixGraph.java
@@ -24,6 +24,7 @@ import ai.grakn.concept.EntityType;
 import ai.grakn.concept.RelationType;
 import ai.grakn.concept.RoleType;
 import ai.grakn.concept.TypeLabel;
+import ai.grakn.test.graphs.TestGraph;
 
 import java.util.function.Consumer;
