+++ grakn-test/src/test/java/ai/grakn/graphs/MovieGraph.java
@@ -28,6 +28,7 @@ import ai.grakn.concept.RoleType;
 import ai.grakn.concept.Rule;
 import ai.grakn.concept.RuleType;
 import ai.grakn.graql.Pattern;
+import ai.grakn.test.graphs.TestGraph;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
