+++ grakn-test/src/test/java/ai/grakn/graphs/TailRecursionGraph.java
@@ -23,6 +23,7 @@ import ai.grakn.concept.EntityType;
 import ai.grakn.concept.RelationType;
 import ai.grakn.concept.RoleType;
 import ai.grakn.concept.TypeLabel;
+import ai.grakn.test.graphs.TestGraph;
 
 import java.util.function.Consumer;
