+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/inference/AbstractInferenceTest.java
@@ -20,8 +20,9 @@ package ai.grakn.test.graql.reasoner.inference;
 
 import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.QueryBuilder;
-import ai.grakn.graphs.AbstractGraph;
 import ai.grakn.test.GraphContext;
+import ai.grakn.test.graphs.AbstractGraph;
+
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Ignore;
